added a new set of queries for SQLServer
